feat: moved all table state logic to dedicated controller
