+++ package.json
@@ -36,7 +36,7 @@
     "build": "tsup",
     "start": "web-dev-server --node-resolve",
     "start:watch": "web-dev-server --node-resolve --watch",
-    "test": "vitest run", 
+    "test": "vitest run",
     "test:watch": "vitest",
     "test:ui": "vitest --ui",
     "preversion": "npm run format && npm test",
@@ -69,6 +69,7 @@
     "typedoc": "^0.28.13",
     "typescript": "^5.8.3",
     "typescript-eslint": "^8.33.1",
+    "vitest": "^4.0.18",
     "vitest-browser-lit": "^1.0.1"
   },
   "dependencies": {
