+++ src/index.ts
@@ -3,6 +3,7 @@ export * from './yatl-button-group';
 export * from './yatl-dropdown';
 export * from './yatl-dropdown-item';
 export * from './yatl-table';
+export * from './yatl-table-controller';
 export * from './yatl-table-ui';
 export * from './yatl-toolbar';
