+++ src/yatl-table-controller/index.ts
@@ -1,0 +1,1 @@
+export * from './yatl-table-controller';
