+++ src/yatl-table/yatl-table.test.ts
@@ -1,5 +1,5 @@
 import { test, expect, vi, describe } from 'vitest';
-import { page, userEvent } from '@vitest/browser/context';
+import { page, userEvent } from 'vitest/browser';
 import { YatlTable } from './yatl-table';
 import './yatl-table';
 import { ColumnOptions } from '../types';
